plugins/google-calendar/requirements.txt
google-api-python-client>=2.0.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=0.4.0
python-dotenv>=0.15.0
pytz>=2021.1 